training_pipeline/requirements.txt
kfp
google-cloud-aiplatform
pandas
scikit-learn
xgboost
numpy
google-cloud-storage
requests
